pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "forge"
version = "0.1.0"
description = "Build & deploy ROS2 in Docker"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-on-whales>=0.59.0",
    "PyYAML>=6.0",
    "Jinja2>=3.1.0",
    "click>=8.0.0",
    "colorama>=0.4.6",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

[project.scripts]
forge = "cli:main"

[tool.setuptools]
py-modules = ["cli"]

[tool.setuptools.packages.find]
where = ["."]
include = ["commands*", "core*", "utils*", "templates*"]

[tool.setuptools.package-data]
templates = ["*.j2"]
